refactor: WarehouseMapper 인터페이스 위치 이동
