+++ src/main/resources/mappers/warehouse/WarehouseMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ssg.meowwms.mapper.WarehouseMapper">
+<mapper namespace="com.ssg.meowwms.mapper.warehouse.WarehouseMapper">
     <!--    창고 등록-->
     <select id="insert">
         insert into Warehouse(name, category_id, manager_id, volume, latitude, longitude)
